requirements.txt
# Voice Note Processor - Python Dependencies
# Install with: pip install -r requirements.txt

# Core dependencies
openai-whisper>=20231117                  # Solo si usarás Whisper localmente también
google-generativeai>=0.3.0                # Gemini API

# Audio processing
torch>=1.9.0
torchaudio>=0.9.0
ffmpeg-python>=0.2.0
pydub>=0.25.1                             # Para cortar audio en fragmentos

# Data processing
numpy>=1.21.0
scipy>=1.7.0

# HTTP and async
aiohttp>=3.9.0                            # Para transcripción paralela con Hugging Face

# JSON and NLP
transformers>=4.19.0

# Utilities
pathlib2>=2.3.0; python_version < "3.4"

# (Opcional) dotenv si usas .env en algún momento
# python-dotenv>=1.0.0